+++ app-specific-code/src/main/java/org/oskari/example/up/ResultsUPHandler.java
@@ -1,5 +1,6 @@
-package org.oskari.example;
+package org.oskari.example.up;
 
+import org.oskari.example.*;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.PreparedStatement;
